Small managers refactoring
