+++ bloss1/src/managers/texture_manager.cpp
@@ -1,16 +1,11 @@
 #include "managers/texture_manager.hpp"
-#include "renderer/opengl/texture.hpp"
+#include "renderer/texture.hpp"
 
 namespace bls
 {
-    std::shared_ptr<Texture> TextureManager::load(const str& name, const str& path, TextureType texture_type)
+    void TextureManager::load(const str& name, std::shared_ptr<Texture> texture)
     {
-        if (exists(name))
-            return textures[name];
-
-        // @TODO: for now, just opengl
-        textures[name] = std::make_shared<OpenGLTexture>(path, texture_type);
-        return textures[name];
+        textures[name] = texture;
     }
 
     std::shared_ptr<Texture> TextureManager::get_texture(const str& name)
