+++ bloss1/src/managers/texture_manager.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 /**
- * @brief @TODO: will be refactored.
+ * @brief Creates, loads and deletes textures.
  */
 
 #include "renderer/texture.hpp"
@@ -11,7 +11,7 @@ namespace bls
     class TextureManager
     {
         public:
-            std::shared_ptr<Texture> load(const str& name, const str& path, TextureType texture_type);
+            void load(const str& name, std::shared_ptr<Texture> texture);
             std::shared_ptr<Texture> get_texture(const str& name);
             bool exists(const str& name);
